Added data
Split into five files due to 25Mb/file limit, something I never thought I'd need to worry about. Contains every one of the 1701935 tabulated knots with 16 or fewer crossings; for each, the name, Gauss code, and Wirtinger number are provided, as well as the existence or nonexistence of a homomorphism to a finite coxeter group, and specifically the existence/nonexistence of a homomorphism to an S, D, and H group. If there is a homomorphism, the seed-strand-to-group-generator mapping producing it is provided. A summary file of statistics from the aggregated data is also provided.